Break up quick check script in smaller composable functions
