+++ src/dev/quick_checks/config.ts
@@ -1,0 +1,22 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { availableParallelism } from 'os';
+import { join } from 'path';
+import { REPO_ROOT } from '@kbn/repo-info';
+
+export const MAX_PARALLELISM = availableParallelism();
+export const IS_CI = process.env.CI === 'true';
+
+export const BUILDKITE_CHECKS_FOLDER = join('.buildkite', 'scripts', 'steps', 'checks');
+export const QUICK_CHECKS_LIST = join(BUILDKITE_CHECKS_FOLDER, 'quick_checks.json');
+export const COLLECT_COMMITS_MARKER_FILE = join(REPO_ROOT, '.collect_commits_marker');
+
+// Maximum output to capture per check (to avoid memory issues)
+export const MAX_OUTPUT_SIZE = 100 * 1024; // 100KB per check
